CORE-16803: Refactor TokenCacheEventProcessorFactory as an OSGi component.
