+++ components/ledger/ledger-utxo-token-cache/src/main/kotlin/net/corda/ledger/utxo/token/cache/factories/TokenCacheEventProcessorFactoryImpl.kt
@@ -3,22 +3,80 @@ package net.corda.ledger.utxo.token.cache.factories
 import net.corda.data.ledger.utxo.token.selection.event.TokenPoolCacheEvent
 import net.corda.data.ledger.utxo.token.selection.key.TokenPoolCacheKey
 import net.corda.data.ledger.utxo.token.selection.state.TokenPoolCacheState
-import net.corda.ledger.utxo.token.cache.converters.EntityConverter
-import net.corda.ledger.utxo.token.cache.converters.EventConverter
+import net.corda.db.connection.manager.DbConnectionManager
+import net.corda.flow.external.events.responses.factory.ExternalEventResponseFactory
+import net.corda.ledger.utxo.token.cache.converters.EntityConverterImpl
+import net.corda.ledger.utxo.token.cache.converters.EventConverterImpl
 import net.corda.ledger.utxo.token.cache.entities.TokenEvent
-import net.corda.ledger.utxo.token.cache.entities.TokenPoolCache
+import net.corda.ledger.utxo.token.cache.entities.internal.TokenPoolCacheImpl
+import net.corda.ledger.utxo.token.cache.handlers.TokenBalanceQueryEventHandler
+import net.corda.ledger.utxo.token.cache.handlers.TokenClaimQueryEventHandler
+import net.corda.ledger.utxo.token.cache.handlers.TokenClaimReleaseEventHandler
 import net.corda.ledger.utxo.token.cache.handlers.TokenEventHandler
+import net.corda.ledger.utxo.token.cache.handlers.TokenLedgerChangeEventHandler
+import net.corda.ledger.utxo.token.cache.queries.impl.SqlQueryProviderTokens
+import net.corda.ledger.utxo.token.cache.repositories.impl.UtxoTokenRepositoryImpl
+import net.corda.ledger.utxo.token.cache.services.ServiceConfiguration
+import net.corda.ledger.utxo.token.cache.services.SimpleTokenFilterStrategy
 import net.corda.ledger.utxo.token.cache.services.TokenCacheEventProcessor
+import net.corda.ledger.utxo.token.cache.services.internal.AvailableTokenServiceImpl
 import net.corda.messaging.api.processor.StateAndEventProcessor
+import net.corda.orm.JpaEntitiesRegistry
+import net.corda.virtualnode.read.VirtualNodeInfoReadService
+import org.osgi.service.component.annotations.Activate
+import org.osgi.service.component.annotations.Component
+import org.osgi.service.component.annotations.Reference
 
-class TokenCacheEventProcessorFactoryImpl constructor(
-    private val eventConverter: EventConverter,
-    private val entityConverter: EntityConverter,
-    private val poolCache: TokenPoolCache,
-    private val tokenCacheEventHandlerMap: Map<Class<*>, TokenEventHandler<in TokenEvent>>
+@Suppress("LongParameterList", "Unused")
+@Component(service = [ TokenCacheEventProcessorFactory::class ])
+class TokenCacheEventProcessorFactoryImpl @Activate constructor(
+    @Reference
+    private val serviceConfiguration: ServiceConfiguration,
+    @Reference
+    private val externalEventResponseFactory: ExternalEventResponseFactory,
+    @Reference
+    private val virtualNodeInfoService: VirtualNodeInfoReadService,
+    @Reference
+    private val dbConnectionManager: DbConnectionManager,
+    @Reference
+    private val jpaEntitiesRegistry: JpaEntitiesRegistry
 ) : TokenCacheEventProcessorFactory {
 
     override fun create(): StateAndEventProcessor<TokenPoolCacheKey, TokenPoolCacheState, TokenPoolCacheEvent> {
-        return TokenCacheEventProcessor(eventConverter, entityConverter,poolCache, tokenCacheEventHandlerMap)
+        val entityConverter = EntityConverterImpl()
+        val eventConverter = EventConverterImpl(entityConverter)
+        val recordFactory = RecordFactoryImpl(externalEventResponseFactory)
+        val tokenFilterStrategy = SimpleTokenFilterStrategy()
+        val sqlQueryProvider = SqlQueryProviderTokens()
+        val utxoTokenRepository = UtxoTokenRepositoryImpl(sqlQueryProvider)
+        val tokenPoolCache = TokenPoolCacheImpl()
+        val availableTokenService = AvailableTokenServiceImpl(
+            virtualNodeInfoService,
+            dbConnectionManager,
+            jpaEntitiesRegistry,
+            utxoTokenRepository,
+            serviceConfiguration
+        )
+
+        val eventHandlerMap = mapOf<Class<*>, TokenEventHandler<in TokenEvent>>(
+            createHandler(
+                TokenClaimQueryEventHandler(
+                    tokenFilterStrategy,
+                    recordFactory,
+                    availableTokenService
+                )
+            ),
+            createHandler(TokenClaimReleaseEventHandler(recordFactory)),
+            createHandler(TokenLedgerChangeEventHandler()),
+            createHandler(TokenBalanceQueryEventHandler(recordFactory, availableTokenService)),
+        )
+        return TokenCacheEventProcessor(eventConverter, entityConverter, tokenPoolCache, eventHandlerMap)
+    }
+
+    private inline fun <reified T : TokenEvent> createHandler(
+        handler: TokenEventHandler<in T>
+    ): Pair<Class<T>, TokenEventHandler<in TokenEvent>> {
+        @Suppress("unchecked_cast")
+        return Pair(T::class.java, handler as TokenEventHandler<in TokenEvent>)
     }
 }
